Correções nas tabelas

+++ database/migrations/2023_08_10_174801_create_proprietarios_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('proprietarios', function (Blueprint $table) {
 
-            $table->increments('id');          // Chave primária autoincrementável do tipo inteiro
+            $table->bigIncrements('id');          // Chave primária autoincrementável do tipo inteiro
             $table->string('nome', 255);            // Coluna do 'nome'
             $table->string('foto', 255);            // Coluna da 'foto'
             $table->string('email', 255);           // Coluna do 'email'
